added modules

+++ src/Etudiant/supprimer_etudiant.php
@@ -1,9 +1,8 @@
 <?php
-include '../connection.php';
-$codeapoge=$_GET["id"];
-echo $codeapoge;
-
-mysqli_query($conn , "DELETE FROM etudiant where code_apoge=$codeapoge");
-
-header('location: ../Etudiants.php?etudiant=deleted');
+	include '../connection.php';
+	if(isset($_GET["id"])){
+		$codeapoge=$_GET["id"];
+		mysqli_query($conn , "DELETE FROM etudiant where code_apoge=$codeapoge");
+		header('location: ../Etudiants.php?etudiant=deleted');
+	}
 ?>
